Spotify: improved check_new_songs()
Improved reliability of the check_new_songs() method: now it compares albums and singles albums by their name to check if there is any new release.
This is better than comparing the whole album dictionary because any album detail edit by the artist could lead to a false new release.
Use names is better than IDs because Spotify makes a lot of multiple instances of the same album for marketing purposes, duplicates are removed by the get_discography() method but we can't determine which album ID is returned in case of a duplicate.
So using the album name seems reasonable: no artist releases two album with the same name